requirements.txt
PyGithub
click
